services/control-api/pyproject.toml
[project]
name = "qaongdur-control-api"
version = "0.1.0"
description = "Qaongdur control plane API with Keycloak token validation."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.7.0",
  "PyJWT[crypto]>=2.10.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
qaongdur-control-api = "control_api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/control_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
